[CELEBORN-1109] Cache RegisterShuffleResponse to improve the processing speed of LifecycleManager
### What changes were proposed in this pull request?
Cache RegisterShuffleResponse to improve the processing speed of LifecycleManager

### Why are the changes needed?
During the processing of the registerShuffle request, constructing the RegisterShuffleResponse instance and serialization can indeed consume a significant amount of time.  When there are a large number of registerShuffle requests that need to be processed by the LifecycleManager simultaneously, the response time of the LifecycleManager will be delayed. Therefore, caching is needed to improve the processing performance of the LifecycleManager.

![image](https://github.com/apache/incubator-celeborn/assets/107825064/06d3cb3c-156a-46c7-a08d-fefa18b26e40)

### Does this PR introduce _any_ user-facing change?

### How was this patch tested?

Closes #2070 from kerwin-zk/issue-1109.

Authored-by: xiyu.zk <xiyu.zk@alibaba-inc.com>
Signed-off-by: zky.zhoukeyong <zky.zhoukeyong@alibaba-inc.com>
(cherry picked from commit ffbbe257fbb41b27638cfc6ee03201f7f6cc6112)
Signed-off-by: zky.zhoukeyong <zky.zhoukeyong@alibaba-inc.com>