[CELEBORN-1126] Set kubernetes resources field for master and worker init container for helm chart
### What changes were proposed in this pull request?
For the `helm` chart. I specified the resources field for the `initContainers` for `worker` and `master` statefulsets.
I used the same values which are specified for the "main" container

### Why are the changes needed?
For users that have a `ResourceQuota` (such as myself), worker and master pods do not start since the initContainers for the statefulsets do not specify the resources (cpu/ memory requests and limits).

### Does this PR introduce _any_ user-facing change?

### [Issue](https://github.com/apache/incubator-celeborn/issues/2094)

### How was this patch tested?
I installed the chart on my GKE cluster.

Closes #2093 from marwansalem/set-k8s-resources-for-init-containers.

Authored-by: Marwan Salem <marwan.saad@incorta.com>
Signed-off-by: Shuang <lvshuang.tb@gmail.com>