[CELEBORN-1149] Improve replica selection when rack aware
### What changes were proposed in this pull request?

As detailed in the jira, this improves replica selection when rack awareness is enabled.
There are primarily two changes:

* Change the distribution of workers in the input workers list so that there is maximum rack diversity between elements in the list.
* Change round robin implementation to independently select replica - instead of based on primary.

### Why are the changes needed?

In our productionization testing environment, we saw disproportionate amount of traffic to a small set of nodes when rack awareness was enabled - which resulted in bringing down the overall throughput of the entire deployment - as all celeborn traffic gets bottlenecked on a small set of nodes.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

This was extensively tested in our environment and validated to improve the overall performance. Note, this is againts a modified version of 0.3, but the specific changes touch parts of master which have not been updated since.
In addition unit tests pass as well.

Closes #2126 from mridulm/rack-aware-replica-selection-improvement.

Lead-authored-by: Mridul Muralidharan <mridul@gmail.com>
Co-authored-by: Mridul Muralidharan <mridulatgmail.com>
Signed-off-by: zky.zhoukeyong <zky.zhoukeyong@alibaba-inc.com>