[CELEBORN-1193] ResettableSlidingWindowReservoir should reset `full` to false
### What changes were proposed in this pull request?
When ResettableSlidingWindowReservoir reset,  it should reset` full` to `false`, `index` to `0`

### Why are the changes needed?
The ResettableSlidingWindowReservoir class, after invoking the reset operation, resets the data to zero, but fails to reset the 'index' and 'full' variables. Consequently, when retrieving a snapshot in the next operation, it is possible to obtain a considerable amount of zeros. This issue extends to the inaccurate calculation of metrics such as average and minimum values.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
add uts

Closes #2182 from lyy-pineapple/slide-bug.

Lead-authored-by: liangyongyuan <2081248500@qq.com>
Co-authored-by: liangyongyuan <liangyongyuan@xiaomi.com>
Signed-off-by: zky.zhoukeyong <zky.zhoukeyong@alibaba-inc.com>