[CELEBORN-1210] Fix potential memory leak in PartitionFilesCleaner
### What changes were proposed in this pull request?

1. let `cleanupExpiredShuffleKey` holds the same lock with add and cleaner thread
2. fix the removal during the iteration
3. simply call `condition.await()` instead of while loop with `condition.await(500ms)`

### Why are the changes needed?

The usage of `LinkedBlockingQueue queue` in `PartitionFilesCleaner` is not a typical producer-consumer model, but an order-agnostic buffer, which confuses me on the first round code reading.

Though `LinkedBlockingQueue` is a thread-safe collection, it means won't get a `ConcurrentModificationException` if the queue is modified while it is iterating, but the iteration is not guaranteed to see all queue entries. As `cleanupExpiredShuffleKey` is not guarded by the lock, elements removal from the cleaner thread may break the iteration and eventually cause a memory leak.

Ref: https://stackoverflow.com/questions/37945981/concurrently-iterating-over-a-blockingqueue

Another issue is the removal inside iteration. The typical usage is
```
Iterator itr = collection.listIterator();
while (itr.hasNext()) {
    if (itr.next() xxx condition) {
        itr.remove();
    }
}
```
The keypoint is that `itr.remove()` should be called instead of `collection.remove(x)`

### Does this PR introduce _any_ user-facing change?

I suppose there is a potential memory leak issue on the Worker without this patch.

### How was this patch tested?

Pass GA.

Closes #2205 from pan3793/CELEBORN-1210.

Authored-by: Cheng Pan <chengpan@apache.org>
Signed-off-by: Cheng Pan <chengpan@apache.org>