[CELEBORN-1253] Improve exception message of fetching chunk failure for WorkerPartitionReader
### What changes were proposed in this pull request?

Improve exception message of fetching chunk failure for `WorkerPartitionReader` including shuffle key.

### Why are the changes needed?

The exception message of fetching chunk failure for `WorkerPartitionReader` does not contain shuffle key of the chunk, which could not match the log of Worker to troubleshooting. It's recommend to add shuffle key in exception message of fetching chunk failure.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

No.

Closes #2261 from SteNicholas/CELEBORN-1253.

Authored-by: SteNicholas <programgeek@163.com>
Signed-off-by: SteNicholas <programgeek@163.com>